refactor(manager): ゲーム一覧の version-up/delete 重ロジックを service に抽出 (一覧WPF化 ①)
ゲーム一覧画面 WPF 化 (#245 modernization) の前段。GameSectionPanel に直書きされていた
バージョンアップ (~300行) と削除 (~200行) のオーケストレーション (コピー/atomic move/
rollback/並行Manager競合/多段削除) を GameVersionUpService / GameDeletionService に抽出
(CLAUDE.md「UI は薄く、ロジックは外へ」)。挙動は完全保存=verbatim move で this.FindForm()
→owner(IWin32Window) / LoadGames()→reload コールバック のみ置換。選択検証 + SessionConflictHelper
は UI 層 (panel) に残す。HandleVersionDirMoveFailure も version service へ移動。

これで ② の WPF GameListPage が同 service を呼ぶだけの薄いページになる (god-file 再生産を回避)。
版数 bump 無し (純粋リファクタ・挙動不変)。bump は ② で。GameSectionPanel: -490/+9 行。
